% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshtmR-package.R
\docType{package}
\name{lshtmR-package}
\alias{lshtmR}
\alias{lshtmR-package}
\title{lshtmR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csthiago/lshtmR}
  \item Report bugs at \url{https://github.com/csthiago/lshtmR/issues}
}

}
\author{
\strong{Maintainer}: Thiago Cerqueira-Silva \email{first.last@example.com}

}
\keyword{internal}
