% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshtmR-package.R
\docType{package}
\name{lshtmR-package}
\alias{lshtmR}
\alias{lshtmR-package}
\title{lshtmR: Mantel-Haenszel procedures for LSHTM classes}
\description{
Mantel-Haenszel for educational purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csthiago/lshtmR}
  \item Report bugs at \url{https://github.com/csthiago/lshtmR/issues}
}

}
\author{
\strong{Maintainer}: Thiago Cerqueira-Silva \email{thiago.silva@lshtm.ac.uk}

}
\keyword{internal}
