% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_fun.R
\name{mh_or_trend}
\alias{mh_or_trend}
\title{Mantel-Haenszel Trend Test for Odds Ratios (internal)}
\usage{
mh_or_trend(data, exposure, outcome, strata_vars = NULL, conf.level = 0.95)
}
\arguments{
\item{data}{A data frame.}

\item{exposure}{Name of numeric exposure variable.}

\item{outcome}{Name of binary outcome variable.}

\item{strata_vars}{Name of stratification variable(s) (currently not used for trend).}

\item{conf.level}{Confidence level.}
}
\value{
A list with trend test results.
}
\description{
Calculates a score test for trend in odds ratios across ordered exposure levels.
This is called internally by mh_or when the exposure has more than 2 levels.
Based on the Mantel extension test for trend. Matches Stata's mhodds command.
}
\keyword{internal}
